Decoupling ComplexYOLOv4 files
