+++ src/perception/object_detector_3d/object_detector_3d/complex_yolov4/README.md
@@ -1,0 +1,5 @@
+# Complex YOLOv4
+
+The PyTorch Implementation based on YOLOv4 of the paper: [Complex-YOLO: Real-time 3D Object Detection on Point Clouds](https://arxiv.org/pdf/1803.06199.pdf)
+
+Most files are taken (some modified) from the GitHub Repo: [https://github.com/maudzung/Complex-YOLOv4-Pytorch/tree/master](https://github.com/maudzung/Complex-YOLOv4-Pytorch/tree/master)
