+++ src/perception/object_detector_3d/object_detector_3d/complex_yolov4/utils/evaluation_utils.py
@@ -1,0 +1,114 @@
+import sys
+import torch
+import numpy as np
+from shapely.geometry import Polygon
+
+sys.path.append('../')
+
+from object_detector_3d.complex_yolov4.utils.kitti_bev_utils import get_corners
+
+
+def cvt_box_2_polygon(box):
+    """
+    :param box: an array of shape [4, 2]
+    :return: a shapely.geometry.Polygon object
+    """
+    # use .buffer(0) to fix a line polygon
+    # more infor: https://stackoverflow.com/questions/13062334/polygon-intersection-error-in-shapely-shapely-geos-topologicalerror-the-opera
+    return Polygon([(box[i, 0], box[i, 1]) for i in range(len(box))]).buffer(0)
+
+
+def iou_rotated_single_vs_multi_boxes_cpu(single_box, multi_boxes):
+    """
+    :param pred_box: Numpy array
+    :param target_boxes: Numpy array
+    :return:
+    """
+
+    s_x, s_y, s_w, s_l, s_im, s_re = single_box
+    s_area = s_w * s_l
+    s_yaw = np.arctan2(s_im, s_re)
+    s_conners = get_corners(s_x, s_y, s_w, s_l, s_yaw)
+    s_polygon = cvt_box_2_polygon(s_conners)
+
+    m_x, m_y, m_w, m_l, m_im, m_re = multi_boxes.transpose(1, 0)
+    targets_areas = m_w * m_l
+    m_yaw = np.arctan2(m_im, m_re)
+    m_boxes_conners = get_corners_vectorize(m_x, m_y, m_w, m_l, m_yaw)
+    m_boxes_polygons = [cvt_box_2_polygon(box_) for box_ in m_boxes_conners]
+
+    ious = []
+    for m_idx in range(multi_boxes.shape[0]):
+        intersection = s_polygon.intersection(m_boxes_polygons[m_idx]).area
+        iou_ = intersection / (s_area + targets_areas[m_idx] - intersection + 1e-16)
+        ious.append(iou_)
+
+    return torch.tensor(ious, dtype=torch.float)
+
+
+def get_corners_vectorize(x, y, w, l, yaw):
+    """bev image coordinates format - vectorization
+
+    :param x, y, w, l, yaw: [num_boxes,]
+    :return: num_boxes x (x,y) of 4 conners
+    """
+    bbox2 = np.zeros((x.shape[0], 4, 2), dtype=np.float32)
+    cos_yaw = np.cos(yaw)
+    sin_yaw = np.sin(yaw)
+
+    # front left
+    bbox2[:, 0, 0] = x - w / 2 * cos_yaw - l / 2 * sin_yaw
+    bbox2[:, 0, 1] = y - w / 2 * sin_yaw + l / 2 * cos_yaw
+
+    # rear left
+    bbox2[:, 1, 0] = x - w / 2 * cos_yaw + l / 2 * sin_yaw
+    bbox2[:, 1, 1] = y - w / 2 * sin_yaw - l / 2 * cos_yaw
+
+    # rear right
+    bbox2[:, 2, 0] = x + w / 2 * cos_yaw + l / 2 * sin_yaw
+    bbox2[:, 2, 1] = y + w / 2 * sin_yaw - l / 2 * cos_yaw
+
+    # front right
+    bbox2[:, 3, 0] = x + w / 2 * cos_yaw - l / 2 * sin_yaw
+    bbox2[:, 3, 1] = y + w / 2 * sin_yaw + l / 2 * cos_yaw
+
+    return bbox2
+
+
+def post_processing_v2(prediction, conf_thresh=0.95, nms_thresh=0.4):
+    """
+        Removes detections with lower object confidence score than 'conf_thres' and performs
+        Non-Maximum Suppression to further filter detections.
+        Returns detections with shape:
+            (x, y, w, l, im, re, object_conf, class_score, class_pred)
+    """
+    output = [None for _ in range(len(prediction))]
+    for image_i, image_pred in enumerate(prediction):
+        # Filter out confidence scores below threshold
+        image_pred = image_pred[image_pred[:, 6] >= conf_thresh]
+        # If none are remaining => process next image
+        if not image_pred.size(0):
+            continue
+        # Object confidence times class confidence
+        score = image_pred[:, 6] * image_pred[:, 7:].max(dim=1)[0]
+        # Sort by it
+        image_pred = image_pred[(-score).argsort()]
+        class_confs, class_preds = image_pred[:, 7:].max(dim=1, keepdim=True)
+        detections = torch.cat((image_pred[:, :7].float(), class_confs.float(), class_preds.float()), dim=1)
+        # Perform non-maximum suppression
+        keep_boxes = []
+        while detections.size(0):
+            # large_overlap = rotated_bbox_iou(detections[0, :6].unsqueeze(0), detections[:, :6], 1.0, False) > nms_thres # not working
+            large_overlap = iou_rotated_single_vs_multi_boxes_cpu(detections[0, :6], detections[:, :6]) > nms_thresh
+            label_match = detections[0, -1] == detections[:, -1]
+            # Indices of boxes with lower confidence scores, large IOUs and matching labels
+            invalid = large_overlap & label_match
+            weights = detections[invalid, 6:7]
+            # Merge overlapping bboxes by order of confidence
+            detections[0, :6] = (weights * detections[invalid, :6]).sum(0) / weights.sum()
+            keep_boxes += [detections[0]]
+            detections = detections[~invalid]
+        if len(keep_boxes) > 0:
+            output[image_i] = torch.stack(keep_boxes)
+
+    return output
